requirements.txt
mysql-connector-python==8.0.33
matplotlib==3.7.1
fpdf2==2.7.4